Dealiasing is now done consistently throughout the code.
BY default (dealias = 0) the maximum wavenumber kmax is set to nx/3 and all wave modes with the magnitude of the wave vector higher than kmax are purged.  This is the most agressive and the most wasteful dealiasing method, but it retains the isotropic features for all carried wavenumbers.  

The purging is done hree times:
1) in the beginning, by manually calling dealias_all from the main program
2) in rhs_velocity it's done implicitly by zeroing the RHS for all |k|>kmax
3) in rhs_scalar it's done in the similar manner.