Changed the dealiasing again.
Now if dealias=0, we use the 2/3 rule: the code keeps the Fourier modes that are inside the cube with the sides (-N/3, N/3].  The rest of the modes in the big cube (-N/2, N/2] are set to zero.  The spectral statistics are calculated using only the modes with |k| \leq k_{max} = N/3.  This is done to reduce the anisotropy of the statistics.