dealias=1 for velocities only.
Inttroduced array ialias(:,:,:) in x_fftw.f90, which is integer*1 and contains the number of possibly aliased wavenumbers at the point (i,j,k) in Fourier space.  ialias=0 inside the cube [-nx/3,nx/3]^3 because tere will be no aliased products.  ialias=1 in teh areas when one of the component of the wave vector is greater than nx/3, ialias=2 where two of these components are greater than nx/3, and ialias=3 where all three components of the wave vector have absolute value larger than nx/3.  This enables us to select the nodes that we need to null in the Fourier space using operators such as "if (ialias(i,j,k).lt.2)".  This array is used in rhs_velocities.f90 now.

Changed logical bug in begin_new: even if the n_scalars=0 but TSCALAR=0.0, int_scalar was .true. and the scalars were carried around (not much damage was done, though because rhs_scalars would not do any actual calculations).

As of now, if we try to call rhs_scalars with dealias=1, it stops the code with error "dealias=1 is not implemented for scalars".

New definition of kmax: if dealias=1, kmax=nx/3*sqrt(2).  if dealias=0, kmax = nx/3, as it was before.

Some debugging messages in m_work, I guess...