Restart files now contain fields in Fourier form (as opposed to X-space form that has been used earlier).  This is done to retain the accuracy in representation of the fields.  The Fourier transform, however accurate, introduces some roundoff errors in the field, in particular, the modes that are set to zero in de-aliasing routines become non-zero, and cause the formal increase in divergence right after the restart.  Besides, we got rid of extra FFT before the writing of the restart file and right after reading it.
Also the restart file now contains DT, the timestep.  This enables the code to start with the same timestep as was used before, and thus run generally faster, because previously the timestep was taken from the .in file.  The user had to set DT deliberatly lower than maximum in order not to overshoot the Courant number.

Some cosmetic changs in main.f90.  Also, now restart read and write is done in parallel (just changed the calls in main.f90)