Implemented rescaling of velocity for decayling isotropic turbulence runs.
The line in the input file that deals with rescales contains two numbers: TRESCALE (when to rescale) and NRESCALE (how many times to rescale).  The rescaling happens at times TRESCALE*1, TRESCALE*2, ..., TRESCALE*NRESCALE.

Rescaling means that the Fourier components of velocity are multipled in such a way that the energy spectrum takes the shape prescribed in the input file by the variables ips_type and peak_wavenum.