Added the les_model=10:
The mixed SGS model for velocities: DSTM + C_mixed * plain Smagorinsky.  The k-equation is closed via the lag-model.  Also the equation for k_sgs is deprecated and not used, because k_sgs = (BT) + (eps T) by definition.

For scalars: The SGS flux model is taken to be the Harlow gradient-diffusivity model (Daly Harlow 10970, Chumakov 2008).

* fixed an annoying bug in init_scalars: the high wavenumber content was initialized but never changed due to the fact that the RHSs were dealiased and did not have any high-wavenumber content.  This led to scalars not decaying properly.

* fixed problem at restart I/O.  Now the restart files do not contain the passive scalars that are not transported.